refactor: move common transforms into common.py and update util methods to have single responsibility
